% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-microbiomeComputations.R
\name{betaDiv}
\alias{betaDiv}
\title{Beta diversity}
\usage{
betaDiv(
  data,
  method = c("bray", "jaccard", "jsd"),
  k = 2,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{AbundanceData object}

\item{method}{string defining the the beta diversity dissimilarity method. Accepted values are 'bray','jaccard', and 'jsd'}

\item{k}{integer determining the number of pcoa axes to return}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns pcoa coordinates calculated from the beta diversity dissimilarity matrix.
}
\examples{
betaDivOutput <- betaDiv(
     getCollection(
	microbiomeData::DiabImmune, 
	"16S (V4) Genus (Relative taxonomic abundance analysis)"), 
     method = "bray", 
     k = 2
)
}
