% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-microbiomeComputations.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(
  data1,
  data2,
  method,
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE),
  ...
)
}
\arguments{
\item{data1}{first dataset. A data.table}

\item{data2}{second dataset. A data.table}

\item{method}{string defining the type of correlation to run.
The currently supported values are specific to the class of data1 and data2.}

\item{format}{string defining the desired format of the result.
The currently supported values are 'data.table' and 'ComputeResult'.}

\item{verbose}{boolean indicating if timed logging is desired}

\item{...}{additional parameters}
}
\value{
data.frame with correlation coefficients or a ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one dataset against variables in a second dataset
}
\examples{
diabImmune_genus <- getCollection(
     microbiomeData::DiabImmune, 
     "16S (V4) Genus", 
     continuousMetadataOnly = TRUE
)

correlationDT <- correlation(
     diabImmune_genus, 
     method = 'spearman', 
     format = 'data.table'
)

correlationOutput <- correlation(
     diabImmune_genus, 
     method = 'spearman', 
     format = 'ComputeResult'
)

alsoCorrelationDT <- getComputeResult(
     correlationOutput, 
     "data.table"
)
}
