% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{correlationNetwork}
\alias{correlationNetwork}
\alias{correlationNetwork,ComputeResult-method}
\alias{correlationNetwork,data.frame-method}
\title{Correlation Network Visualization}
\usage{
correlationNetwork(object)

\S4method{correlationNetwork}{ComputeResult}(object)

\S4method{correlationNetwork}{data.frame}(object)
}
\arguments{
\item{object}{A ComputeResult or data.frame}
}
\description{
Visualize a correlation result as a network
}
