% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-markdown.R
\name{createRMarkdownFromComputeResult}
\alias{createRMarkdownFromComputeResult}
\title{Create R Markdown file from compute result
This function will create an R Markdown file that from a ComputeResult object. The markdown
file will describe how to access the data in the ComputeResult object and create a plot. Finally,
it will render the R Markdown file to HTML. Templates for these mardown files are stored in the
inst/rmarkdown/templates directory.}
\usage{
createRMarkdownFromComputeResult(result, rmd_file_name = "tmp.Rmd")
}
\arguments{
\item{result}{A compute result object}

\item{rmd_file_name}{The name of the R Markdown file to be created. Default is 'tmp.Rmd'}
}
\value{
A list containing the plots created for the R markdown file.
}
\description{
Create R Markdown file from compute result
This function will create an R Markdown file that from a ComputeResult object. The markdown
file will describe how to access the data in the ComputeResult object and create a plot. Finally,
it will render the R Markdown file to HTML. Templates for these mardown files are stored in the
inst/rmarkdown/templates directory.
}
