% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-differentialAbundance.R
\name{differentialAbundance}
\alias{differentialAbundance}
\alias{differentialAbundance,CollectionWithMetadata,character,missingOrNULL,missingOrNULL-method}
\alias{differentialAbundance,CollectionWithMetadata,character,function,missingOrNULL-method}
\alias{differentialAbundance,CollectionWithMetadata,character,function,function-method}
\title{Differential abundance}
\usage{
differentialAbundance(
  data,
  covariate,
  groupAPredicate,
  groupBPredicate,
  method = c("Maaslin2", "DESeq2"),
  verbose = c(TRUE, FALSE)
)

\S4method{differentialAbundance}{CollectionWithMetadata,character,missingOrNULL,missingOrNULL}(
  data,
  covariate,
  groupAPredicate,
  groupBPredicate,
  method = c("Maaslin2", "DESeq2"),
  verbose = c(TRUE, FALSE)
)

\S4method{differentialAbundance}{CollectionWithMetadata,character,function,missingOrNULL}(
  data,
  covariate,
  groupAPredicate,
  groupBPredicate,
  method = c("Maaslin2", "DESeq2"),
  verbose = c(TRUE, FALSE)
)

\S4method{differentialAbundance}{CollectionWithMetadata,character,function,function}(
  data,
  covariate,
  groupAPredicate,
  groupBPredicate,
  method = c("Maaslin2", "DESeq2"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{AbundanceData object}

\item{covariate}{character vector giving the name of a metadata variable of interest}

\item{method}{string defining the the differential abundance method. Accepted values are 'DESeq2' and 'Maaslin2'.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns the fold change and associated p value for
a differential abundance analysis comparing samples in two groups.
It is useful for recreating the results of differential abundance analyses
from MicrobiomeDB.org, but we recognize makes some assumptions about the data
which may not be valid in other contexts. For better support of longitudinal studies
or metabolomic data, for example, please see our wrapper/ helper methods for
Maaslin2 (\code{MicrobiomeDB::Maaslin2}) and DESeq2 (\code{DESeqDataSetFromCollection}).
}
