% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getCollection}
\alias{getCollection}
\alias{getCollection,MbioDataset-method}
\alias{getCollection,MbioDataset,character-method}
\title{Get Microbiome Dataset Collection}
\usage{
getCollection(
  object,
  collectionName,
  format = c("AbundanceData", "phyloseq", "Collection"),
  continuousMetadataOnly = c(FALSE, TRUE)
)

\S4method{getCollection}{MbioDataset}(
  object,
  collectionName = character(0),
  format = c("AbundanceData", "phyloseq", "Collection"),
  continuousMetadataOnly = c(FALSE, TRUE)
)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{collectionName}{The name of the collection to return}

\item{format}{The format of the collection to return. Currently supported options are "AbundanceData", "phyloseq" and "Collection".}

\item{continuousMetadataOnly}{If TRUE, only continuous metadata will be returned. If FALSE, all metadata will be returned.}
}
\value{
An AbundanceData, phyloseq, or Collection object representing the collection and any associated study metadata
}
\description{
Get a collection from the Microbiome Dataset. The collection will be returned
as an AbundanceData, phyloseq, or Collection object.
}
\examples{
genus <- getCollection(
     microbiomeData::DiabImmune, 
     "16S (V4) Genus (Relative taxonomic abundance analysis)"
)

genus_phyloseq <- getCollection(
     microbiomeData::DiabImmune, 
     "16S (V4) Genus (Relative taxonomic abundance analysis)", 
     format = "phyloseq"
)

## to pass to correlation method, we want only continuous metadata
genus_continuous <- getCollection(
     microbiomeData::DiabImmune, 
     "16S (V4) Genus (Relative taxonomic abundance analysis)", 
     continuousMetadataOnly = TRUE
) 

## with no metadata
genus_collection <- getCollection(
     microbiomeData::DiabImmune, 
     "16S (V4) Genus (Relative taxonomic abundance analysis)", 
     format = "Collection"
)
}
