% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getCollectionNames,MbioDataset-method}
\alias{getCollectionNames,MbioDataset-method}
\title{Get Names of Collections}
\usage{
\S4method{getCollectionNames}{MbioDataset}(object)
}
\arguments{
\item{object}{An MbioDataset}
}
\value{
A character vector of collection names
}
\description{
Get the names of the collections in a MbioDataset object
}
