% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getCollectionVariableNames,MbioDataset-method}
\alias{getCollectionVariableNames,MbioDataset-method}
\title{Get Microbiome Dataset Collection Variable Names}
\usage{
\S4method{getCollectionVariableNames}{MbioDataset}(object, collectionName)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{collectionName}{The name of the collection to return the variable names for}
}
\value{
a character vector of the variable names in the requested collection
}
\description{
Get the variable names in a collection in the Microbiome Dataset.
}
\examples{
variableNames <- getCollectionVariableNames(microbiomeData::DiabImmune, "16S (V4) Genus")
}
