% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getComputeResult}
\alias{getComputeResult}
\alias{getComputeResult,ComputeResult-method}
\alias{getComputeResult,CorrelationResult-method}
\alias{getComputeResult,DifferentialAbundanceResult-method}
\title{Get Microbiome Dataset Compute Result}
\usage{
getComputeResult(object, format = c("data.table"), ...)

\S4method{getComputeResult}{ComputeResult}(object, format = c("data.table", "igraph"))

\S4method{getComputeResult}{CorrelationResult}(object, format = c("data.table", "igraph"))

\S4method{getComputeResult}{DifferentialAbundanceResult}(object, format = c("data.table"))
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{format}{The format of the compute result. Currently only "data.table" and "igraph" are supported.}
}
\value{
The compute result in the specified format
}
\description{
Get the compute result from a Microbiome Dataset in a particular format.
Some formats may not be supported for all compute results.
}
