% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getComputeResultWithMetadata}
\alias{getComputeResultWithMetadata}
\title{Get Microbiome Dataset Compute Result With Metadata}
\usage{
getComputeResultWithMetadata(
  object,
  dataset,
  format = c("data.table"),
  metadataVariables = NULL
)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{dataset}{The MbioDataset, AbundanceData or Collection object from which the compute result was obtained.}

\item{format}{The format of the compute result. Currently only "data.table" is supported.}

\item{metadataVariables}{The metadata variables to include in the compute result. If NULL, no metadata variables will be included.}
}
\value{
The compute result in the specified format
}
\description{
Get the compute result from a Microbiome Dataset in a particular format with metadata.
}
