% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getCuratedDatasetNames}
\alias{getCuratedDatasetNames}
\title{List Curated Datasets}
\usage{
getCuratedDatasetNames()
}
\description{
This function lists curated datasets from MicrobiomeDB.org which are available in this package.
}
