% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getMetadataVariableNames,MbioDataset-method}
\alias{getMetadataVariableNames,MbioDataset-method}
\title{Get Variable Names of Metadata}
\usage{
\S4method{getMetadataVariableNames}{MbioDataset}(object)
}
\arguments{
\item{object}{An MbioDataset}
}
\value{
a character vector of metadata variable names
}
\description{
Get the names of the metadata variables in an MbioDataset.
}
