% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getMetadataVariableSummary,MbioDataset-method}
\alias{getMetadataVariableSummary,MbioDataset-method}
\title{Get Summary of Metadata Variables}
\usage{
\S4method{getMetadataVariableSummary}{MbioDataset}(object, variable)
}
\arguments{
\item{object}{An MbioDataset}

\item{variable}{A character vector representing the name of the metadata variable to summarize}
}
\value{
a table summarizing the values of the requested metadata variable
}
\description{
Get a summary of the requested metadata variable in an MbioDataset.
}
\examples{
getMetadataVariableSummary(microbiomeData::DiabImmune, "age_months")
getMetadataVariableSummary(microbiomeData::DiabImmune, "sex")
getMetadataVariableSummary(microbiomeData::DiabImmune, "country")
}
