% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getSampleMetadataIdColumns,MbioDataset-method}
\alias{getSampleMetadataIdColumns,MbioDataset-method}
\title{Get Sample Metadata Id Column Names}
\usage{
\S4method{getSampleMetadataIdColumns}{MbioDataset}(object)
}
\arguments{
\item{object}{MbioDataset}
}
\value{
a character vector of id column names
}
\description{
Get the names of the record and ancestor id columns in the sample metadata of an MbioDataset object.
}
