% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getVariables}
\alias{getVariables}
\alias{getVariables,MbioDataset-method}
\alias{getVariables,MbioDataset,character-method}
\title{Get Microbiome Dataset Variables}
\usage{
getVariables(object, variables)

\S4method{getVariables}{MbioDataset}(object, variables)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{variables}{The names of the variables to return. This should be a named list
where the names are collection names and the values are variable names for that collection.
For the case of metadata variables, the name should be "metadata".}
}
\value{
a data.table of the requested variables
}
\description{
Get the variables in the Microbiome Dataset by their names.
The requested variables could belong to any collection or
to the metadata. The returned data.table will contain the
requested variables as columns and any appropriate identifiers.
If one of the requested variables cannot be returned, a warning
will be printed.
}
\examples{
getCollectionVariableNames(microbiomeData::DiabImmune, "16S (V4) Genus")
getMetadataVariableNames(microbiomeData::DiabImmune)
variablesDT <- getVariables(
     microbiomeData::DiabImmune, 
     list("metadata" = c("age_months", "sex"),
          "16S (V4) Genus" = "Bacteroides", 
          "Shotgun metagenomics Metagenome enzyme pathway abundance data" = "ANAGLYCOLYSIS-PWY: glycolysis III (from glucose)"
     )
)
}
