% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers-MbioDataset.R
\name{importBIOM}
\alias{importBIOM}
\title{Import BIOM data}
\usage{
importBIOM(
  normalizationMethod = c("TSS", "none"),
  keepRawValues = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE),
  ...
)
}
\arguments{
\item{normalizationMethod}{Normalization method to use on they assay data. Options are "none" and "TSS".
Applying TSS normalization to absolute taxonomic abundances produces relative taxonomic abundances. Default is "TSS".}

\item{keepRawValues}{Keep the raw assay values as well as the normalized values.}

\item{verbose}{Print messages}

\item{...}{Arguments to pass to mia::makeTreeSEFromBiom}
}
\value{
A MbioDataset
}
\description{
Import data from BIOM results to MbioDataset. There is
some loss of granularity in this process. It results
in a simpler and more performant object which is compliant
with the MicrobiomeDB infrastructure. See \code{mia::makeTreeSEFromBiom}
for documentation.
}
\examples{
rich_dense_file = system.file("extdata", "rich_dense_otu_table.biom",
                               package = "biomformat")

mbioDataset <- importBIOM(
     normalizationMethod = "none", 
     keepRawValues = TRUE, 
     verbose = TRUE, 
     rich_dense_file
)

rich_dense_biom = biomformat::read_biom(rich_dense_file)

mbioDataset <- importBIOM(
     normalizationMethod = "none", 
     keepRawValues = TRUE, 
     verbose = TRUE, 
     rich_dense_biom
)
}
