% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers-MbioDataset.R
\name{importTreeSummarizedExperiment}
\alias{importTreeSummarizedExperiment}
\alias{importTreeSE}
\title{Import TreeSummarizedExperiment}
\usage{
importTreeSummarizedExperiment(
  data,
  normalizationMethod = c("TSS", "none"),
  keepRawValues = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)

importTreeSE(
  data,
  normalizationMethod = c("TSS", "none"),
  keepRawValues = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{A TreeSummarizedExperiment}

\item{normalizationMethod}{Normalization method to use on they assay data. Options are "none" and "TSS".
Applying TSS normalization to absolute taxonomic abundances produces relative taxonomic abundances. Default is "TSS".}

\item{keepRawValues}{Keep the raw assay values as well as the normalized values.}

\item{verbose}{Print messages}
}
\value{
A MbioDataset
}
\description{
Import data from TreeSummarizedExperiment to MbioDataset.
There is some loss of granularity in this process. It results
in a simpler and more performant object which is compliant
with the MicrobiomeDB infrastructure.
}
