% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-differentialAbundance.R
\name{Maaslin2}
\alias{Maaslin2}
\alias{Maaslin2,CollectionWithMetadata-method}
\title{Maaslin2 from a Collection}
\usage{
Maaslin2(data, verbose = c(TRUE, FALSE), ...)

\S4method{Maaslin2}{CollectionWithMetadata}(data, verbose = c(TRUE, FALSE), ...)
}
\arguments{
\item{data}{a CollectionWithMetadata}

\item{verbose}{boolean indicating if timed logging is desired}

\item{...}{additional arguments to pass to Maaslin2::Maaslin2}
}
\description{
MaAsLin2 finds associations between microbiome
meta-omics features and complex metadata in population-scale
epidemiological studies. The software includes multiple
analysis methods (including support for multiple covariates and repeated measures)
filtering, normalization, and transform options to customize analysis for your specific study.
}
