% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-MbioDataset.R
\name{MbioDataset}
\alias{MbioDataset}
\alias{MbioDataset,missing,missing,missing-method}
\alias{MbioDataset,Collections,SampleMetadata,missing-method}
\alias{MbioDataset,Collections,SampleMetadata,data.frame-method}
\alias{MbioDataset,Collections,SampleMetadata,character-method}
\alias{MbioDataset,Collections,data.frame,missing-method}
\alias{MbioDataset,Collections,data.frame,data.frame-method}
\alias{MbioDataset,Collections,data.frame,character-method}
\alias{MbioDataset,Collections,list,missing-method}
\alias{MbioDataset,Collections,list,data.frame-method}
\alias{MbioDataset,Collections,list,character-method}
\alias{MbioDataset,Collections,missing,missing-method}
\alias{MbioDataset,Collections,missing,data.frame-method}
\alias{MbioDataset,Collections,missing,character-method}
\alias{MbioDataset,Collections,character,missing-method}
\alias{MbioDataset,Collections,character,data.frame-method}
\alias{MbioDataset,Collections,character,character-method}
\alias{MbioDataset,Collection,SampleMetadata,missing-method}
\alias{MbioDataset,Collection,SampleMetadata,data.frame-method}
\alias{MbioDataset,Collection,SampleMetadata,character-method}
\alias{MbioDataset,Collection,data.frame,missing-method}
\alias{MbioDataset,Collection,data.frame,data.frame-method}
\alias{MbioDataset,Collection,data.frame,character-method}
\alias{MbioDataset,Collection,list,missing-method}
\alias{MbioDataset,Collection,list,data.frame-method}
\alias{MbioDataset,Collection,list,character-method}
\alias{MbioDataset,Collection,missing,missing-method}
\alias{MbioDataset,Collection,missing,data.frame-method}
\alias{MbioDataset,Collection,missing,character-method}
\alias{MbioDataset,Collection,character,missing-method}
\alias{MbioDataset,Collection,character,data.frame-method}
\alias{MbioDataset,Collection,character,character-method}
\alias{MbioDataset,list,SampleMetadata,missing-method}
\alias{MbioDataset,list,SampleMetadata,data.frame-method}
\alias{MbioDataset,list,SampleMetadata,character-method}
\alias{MbioDataset,list,data.frame,missing-method}
\alias{MbioDataset,list,data.frame,data.frame-method}
\alias{MbioDataset,list,data.frame,character-method}
\alias{MbioDataset,list,missing,missing-method}
\alias{MbioDataset,list,missing,data.frame-method}
\alias{MbioDataset,list,missing,character-method}
\alias{MbioDataset,list,list,missing-method}
\alias{MbioDataset,list,list,data.frame-method}
\alias{MbioDataset,list,list,character-method}
\alias{MbioDataset,list,character,missing-method}
\alias{MbioDataset,list,character,data.frame-method}
\alias{MbioDataset,list,character,character-method}
\alias{MbioDataset,data.frame,SampleMetadata,missing-method}
\alias{MbioDataset,data.frame,SampleMetadata,data.frame-method}
\alias{MbioDataset,data.frame,SampleMetadata,character-method}
\alias{MbioDataset,data.frame,list,missing-method}
\alias{MbioDataset,data.frame,list,data.frame-method}
\alias{MbioDataset,data.frame,list,character-method}
\alias{MbioDataset,data.frame,missing,missing-method}
\alias{MbioDataset,data.frame,missing,data.frame-method}
\alias{MbioDataset,data.frame,missing,character-method}
\alias{MbioDataset,data.frame,data.frame,missing-method}
\alias{MbioDataset,data.frame,data.frame,data.frame-method}
\alias{MbioDataset,data.frame,data.frame,character-method}
\alias{MbioDataset,data.frame,character,missing-method}
\alias{MbioDataset,data.frame,character,data.frame-method}
\alias{MbioDataset,data.frame,character,character-method}
\alias{MbioDataset,character,character,missing-method}
\alias{MbioDataset,character,character,data.frame-method}
\alias{MbioDataset,character,character,character-method}
\alias{MbioDataset,character,list,missing-method}
\alias{MbioDataset,character,list,data.frame-method}
\alias{MbioDataset,character,list,character-method}
\alias{MbioDataset,character,missing,missing-method}
\alias{MbioDataset,character,missing,data.frame-method}
\alias{MbioDataset,character,missing,character-method}
\alias{MbioDataset,character,data.frame,missing-method}
\alias{MbioDataset,character,data.frame,data.frame-method}
\alias{MbioDataset,character,data.frame,character-method}
\alias{MbioDataset,character,SampleMetadata,missing-method}
\alias{MbioDataset,character,SampleMetadata,data.frame-method}
\alias{MbioDataset,character,SampleMetadata,character-method}
\title{Create a Microbiome Dataset}
\usage{
MbioDataset(collections, metadata, ontology)

\S4method{MbioDataset}{missing,missing,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,SampleMetadata,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,SampleMetadata,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,SampleMetadata,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,data.frame,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,data.frame,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,data.frame,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,list,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,list,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,list,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,missing,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,missing,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,missing,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,character,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,character,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collections,character,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,SampleMetadata,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,SampleMetadata,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,SampleMetadata,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,data.frame,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,data.frame,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,data.frame,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,list,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,list,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,list,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,missing,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,missing,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,missing,character}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,character,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,character,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{Collection,character,character}(collections, metadata, ontology)

\S4method{MbioDataset}{list,SampleMetadata,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{list,SampleMetadata,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{list,SampleMetadata,character}(collections, metadata, ontology)

\S4method{MbioDataset}{list,data.frame,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{list,data.frame,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{list,data.frame,character}(collections, metadata, ontology)

\S4method{MbioDataset}{list,missing,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{list,missing,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{list,missing,character}(collections, metadata, ontology)

\S4method{MbioDataset}{list,list,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{list,list,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{list,list,character}(collections, metadata, ontology)

\S4method{MbioDataset}{list,character,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{list,character,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{list,character,character}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,SampleMetadata,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,SampleMetadata,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,SampleMetadata,character}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,list,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,list,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,list,character}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,missing,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,missing,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,missing,character}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,data.frame,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,data.frame,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,data.frame,character}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,character,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,character,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{data.frame,character,character}(collections, metadata, ontology)

\S4method{MbioDataset}{character,character,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{character,character,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{character,character,character}(collections, metadata, ontology)

\S4method{MbioDataset}{character,list,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{character,list,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{character,list,character}(collections, metadata, ontology)

\S4method{MbioDataset}{character,missing,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{character,missing,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{character,missing,character}(collections, metadata, ontology)

\S4method{MbioDataset}{character,data.frame,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{character,data.frame,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{character,data.frame,character}(collections, metadata, ontology)

\S4method{MbioDataset}{character,SampleMetadata,missing}(collections, metadata, ontology)

\S4method{MbioDataset}{character,SampleMetadata,data.frame}(collections, metadata, ontology)

\S4method{MbioDataset}{character,SampleMetadata,character}(collections, metadata, ontology)
}
\arguments{
\item{collections}{A list of Collection objects, a data.frame containing multiple collections,
or a character vector containing one or more file path(s)}

\item{metadata}{A SampleMetadata object, a data.frame containing sample metadata,
or a list of file path(s)}

\item{ontology}{An data.frame containing the ontology of the dataset, or a character vector
containing a file path to a data.frame}
}
\description{
This is a constructor for the MbioDataset class. It creates a MbioDataset containing
a list of Collections and a SampleMetadata object.
}
