% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-markdown.R
\name{mbiodb_html_format}
\alias{mbiodb_html_format}
\title{Custom HTML template, based on the custom}
\usage{
mbiodb_html_format(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \@code{html_document}}
}
\description{
Loads additional style and template file
}
