% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeDB-package.R
\docType{package}
\name{MicrobiomeDB-package}
\alias{MicrobiomeDB}
\alias{MicrobiomeDB-package}
\title{MicrobiomeDB: Curated Data and Analysis Tools from MicrobiomeDB.org}
\description{
This package is intended to be used to explore the curated datasets from MicrobiomeDB.org, as well as your own datasets. It comes pre-packaged with the same functions used to power the analysis tools from the website. It also contains functions to facilitate easily transforming data between our custom objects, phyloseq objects, and .biom files that you can upload to the website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/microbiomeDB/MicrobiomeDB}
  \item \url{https://microbiomedb.github.io/MicrobiomeDB/}
  \item Report bugs at \url{https://github.com/microbiomeDB/MicrobiomeDB/issues}
}

}
\author{
\strong{Maintainer}: Danielle Callan \email{dcallan@upenn.edu}

}
\keyword{internal}
