% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-microbiomeComputations.R
\name{rankedAbundance}
\alias{rankedAbundance}
\title{Ranked abundance}
\usage{
rankedAbundance(
  data,
  method = c("median", "max", "q3", "variance"),
  cutoff = 10,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{AbundanceData object}

\item{method}{string defining the ranking strategy by which to order the taxa. Accepted values are 'median','max','q3',and 'variance'. Note that taxa that return a value of 0 for a given method will not be included in the results.}

\item{cutoff}{integer indicating the maximium number of taxa to be kept after ranking.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns abundances, ranked by a selected ranking function
}
\examples{
rankedAbundOutput <- rankedAbundance(getCollection(microbiomeData::DiabImmune, "16S (V4) Genus"), method = "median")
}
