% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-microbiomeComputations.R
\name{selfCorrelation}
\alias{selfCorrelation}
\title{Self Correlation}
\usage{
selfCorrelation(
  data,
  method = c("spearman", "pearson", "sparcc"),
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE),
  ...
)
}
\arguments{
\item{data}{An AbundanceData object}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman','pearson' and 'sparcc'}

\item{format}{string defining the desired format of the result. The currently supported values are 'data.table' and 'ComputeResult'.}

\item{verbose}{boolean indicating if timed logging is desired}

\item{...}{additional parameters}
}
\value{
ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one AbundanceData object against itself. It generally serves as a
convenience wrapper around veupathUtils::correlation, with the exception that it additionally supports sparcc.
}
\examples{
correlationDT <- selfCorrelation(getCollection(microbiomeData::DiabImmune, "16S (V4) Genus"), method = 'sparcc', format = 'data.table')
correlationOutput <- selfCorrelation(getCollection(microbiomeData::DiabImmune, "16S (V4) Genus"), method = 'sparcc', format = 'ComputeResult')
alsoCorrelationDT <- getComputeResult(correlationOutput, "data.table")
}
