% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ComputeResult.R
\name{correlationNetwork}
\alias{correlationNetwork}
\alias{correlationNetwork,ComputeResult-method}
\alias{correlationNetwork,data.frame-method}
\title{Correlation Network Visualization}
\usage{
correlationNetwork(
  object,
  correlationCoefThreshold = 0.5,
  pValueThreshold = 0.05,
  ...
)

\S4method{correlationNetwork}{ComputeResult}(
  object,
  correlationCoefThreshold = 0.5,
  pValueThreshold = 0.05,
  ...
)

\S4method{correlationNetwork}{data.frame}(
  object,
  correlationCoefThreshold = 0.5,
  pValueThreshold = 0.05,
  bipartiteNetwork = c(FALSE, TRUE)
)
}
\arguments{
\item{object}{A ComputeResult or data.frame}

\item{correlationCoefThreshold}{threshold to filter edges by correlation coefficient.
Edges with correlation coefficients below this threshold will be removed. Default is .5}

\item{pValueThreshold}{threshold to filter edges by p-value. Edges with p-values above this threshold will be removed. Default is .05}

\item{...}{additional arguments}

\item{bipartiteNetwork}{Should the network use a bipartite or unipartite layout? Defaults to unipartite.}
}
\description{
Visualize a correlation result as a network
}
