% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MbioDataset.R
\name{getCollection}
\alias{getCollection}
\alias{getCollection,MbioDataset-method}
\alias{getCollection,MbioDataset,character-method}
\title{Get Microbiome Dataset Collection}
\usage{
getCollection(
  object,
  collectionName,
  format = c("AbundanceData", "phyloseq", "Collection"),
  continuousMetadataOnly = c(FALSE, TRUE)
)

\S4method{getCollection}{MbioDataset}(
  object,
  collectionName = character(0),
  format = c("AbundanceData", "phyloseq", "Collection"),
  continuousMetadataOnly = c(FALSE, TRUE)
)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{collectionName}{The name of the collection to return}

\item{format}{The format of the collection to return. Currently supported options are "AbundanceData", "phyloseq", and "Collection".}

\item{continuousMetadataOnly}{If TRUE, only continuous metadata will be returned. If FALSE, all metadata will be returned.}
}
\value{
An AbundanceData, phyloseq, or Collection object representing the collection and any associated study metadata
}
\description{
Get a collection from the Microbiome Dataset. The collection will be returned
as an AbundanceData, phyloseq, or Collection object.
}
