% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ComputeResult.R
\name{getComputeResultWithMetadata}
\alias{getComputeResultWithMetadata}
\alias{getComputeResultWithMetadata,ComputeResult,MbioDataset-method}
\alias{getComputeResultWithMetadata,ComputeResult,Collection-method}
\alias{getComputeResultWithMetadata,ComputeResult,AbundanceData-method}
\title{Get Microbiome Dataset Compute Result With Metadata}
\usage{
getComputeResultWithMetadata(
  object,
  dataset,
  format = c("data.table"),
  metadataVariables = NULL
)

\S4method{getComputeResultWithMetadata}{ComputeResult,MbioDataset}(
  object,
  dataset = NULL,
  format = c("data.table"),
  metadataVariables = NULL
)

\S4method{getComputeResultWithMetadata}{ComputeResult,Collection}(
  object,
  dataset = NULL,
  format = c("data.table"),
  metadataVariables = NULL
)

\S4method{getComputeResultWithMetadata}{ComputeResult,AbundanceData}(
  object,
  dataset = NULL,
  format = c("data.table"),
  metadataVariables = NULL
)
}
\arguments{
\item{object}{A Microbiome Dataset}

\item{dataset}{The MbioDataset, AbundanceData or Collection object from which the compute result was obtained.}

\item{format}{The format you want the compute result in. Currently only "data.table" is supported.}

\item{metadataVariables}{The metadata variables to include in the compute result. If NULL, no metadata variables will be included.}
}
\value{
The compute result in the specified format
}
\description{
Get the compute result from a Microbiome Dataset in a particular format with metadata.
}
\examples{
alphaDivOutput <- MicrobiomeDB::alphaDiv(
     getCollection(microbiomeData::DiabImmune, "16S (V4) Genus"), 
     method='shannon', 
     verbose=FALSE
)
alphaDivDT <- getComputeResultWithMetadata(alphaDivOutput, microbiomeData::DiabImmune, metadataVariables = c('country', 'delivery_mode'))
}
