% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rankedAbundance}
\alias{alphaDiv}
\alias{betaDiv}
\alias{correlation}
\alias{selfCorrelation}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{microbiomeComputations}{\code{\link[microbiomeComputations:alphaDiv-methods]{alphaDiv}}, \code{\link[microbiomeComputations:betaDiv-methods]{betaDiv}}, \code{\link[microbiomeComputations:rankedAbundance-methods]{rankedAbundance}}}

  \item{veupathUtils}{\code{\link[veupathUtils]{correlation}}, \code{\link[veupathUtils]{selfCorrelation}}}
}}

