% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computes.R, R/data.R, R/methods-data.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rankedAbundance}
\alias{alphaDiv}
\alias{betaDiv}
\alias{differentialAbundance}
\alias{correlation}
\alias{selfCorrelation}
\alias{DiabImmune}
\alias{FARMM}
\alias{Bangladesh}
\alias{HMP_WGS}
\alias{BONUS}
\alias{NICU_NEC}
\alias{getCollection}
\alias{getCollectionNames}
\alias{updateCollectionName}
\alias{getMetadataVariableNames}
\alias{getSampleMetadata}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{microbiomeComputations}{\code{\link[microbiomeComputations]{alphaDiv}}, \code{\link[microbiomeComputations]{betaDiv}}, \code{\link[microbiomeComputations]{correlation}}, \code{\link[microbiomeComputations]{differentialAbundance}}, \code{\link[microbiomeComputations]{rankedAbundance}}, \code{\link[microbiomeComputations]{selfCorrelation}}}

  \item{microbiomeData}{\code{\link[microbiomeData]{Bangladesh}}, \code{\link[microbiomeData]{BONUS}}, \code{\link[microbiomeData]{DiabImmune}}, \code{\link[microbiomeData]{FARMM}}, \code{\link[microbiomeData]{getCollection}}, \code{\link[microbiomeData]{getCollectionNames}}, \code{\link[microbiomeData]{getMetadataVariableNames}}, \code{\link[microbiomeData]{getSampleMetadata}}, \code{\link[microbiomeData]{HMP_WGS}}, \code{\link[microbiomeData]{NICU_NEC}}, \code{\link[microbiomeData]{updateCollectionName}}}
}}

